requirements.txt
psutil
click
tqdm
imageio
imageio-ffmpeg==0.4.3
pyspng==0.1.0
scipy 
requests
gfpgan>=1.3.5
realesrgan
basicsr>=1.4.2
facexlib>=0.2.5
Pillow
lmdb
opencv-python
pyyaml
tb-nightly
yapf
uvicorn
fastapi
pydantic
pydantic_settings
python-multipart
python-dotenv
aiofiles
httpx
jinja2